Try to identify new errors outside diff lines
Add some logic that uses a heuristic to try to identify errors that were
introduced as a side-effect of the changes in the diff, but outside the
diff lines. For each error that's not on a modified line, check whether
the same error with the same source code and same column offset is
present in the lint results for the old version of the file. Line
numbers are ignored as lines may have been inserted or removed before
the line being checked.